refactor(cave): 将 cave-prompt 并入 larkcave 插件与 cave 指令帮助
- 移除独立插件 nonebot_plugin_cave_image_prompt
- /cave-prompt on|off 开关指令迁入 larkcave 的 commands/prompt.py
- 私聊单图投稿询问消息处理器一并迁入
- CaveImagePromptConfig 模型移入 larkcave/models.py，保留原表名（无需数据迁移）
- cave 指令帮助新增 u8/u9：cave-prompt 用法
- 语言文件合并到 larkcave.yaml（prompt 命名空间）
- 更新根/src pyproject.toml 插件注册，同步调整测试导入
