--- src/lang/zh_hans/cave_image_prompt.yaml
@@ -1,17 +1,0 @@
-ask: |
-  检测到您发送了一张图片！是否要投稿到回声洞？
-  请回复 "是"、"否"、"y" 或 "n"
-"yes": "是"
-"no": "否"
-cancelled: "已取消投稿"
-enabled: "已开启私聊图片投稿询问喵～之后私聊发送单张图片时，会询问是否投稿到回声洞"
-disabled: "已关闭私聊图片投稿询问喵～私聊发图将正常交给 AI 聊天处理"
-status_on: "当前已开启私聊图片投稿询问喵～发送 /cave-prompt off 可关闭"
-status_off: "当前已关闭私聊图片投稿询问喵～发送 /cave-prompt on 可开启"
-usage: "用法：/cave-prompt on|off（不带参数可查看当前状态）"
-help:
-  d: "私聊图片投稿询问开关"
-  i: "私聊发送单张图片时，询问是否投稿到回声洞。默认关闭，用户可自行开启。"
-  u1: "cave-prompt"
-  u2: "cave-prompt on"
-  u3: "cave-prompt off"
