--- src/plugins/nonebot_plugin_cave_image_prompt/help.yaml
@@ -1,10 +1,0 @@
-plugin: cave_image_prompt
-commands:
-  cave-prompt:
-    description: help.d
-    details: help.i
-    usages:
-      - help.u1
-      - help.u2
-      - help.u3
-    category: setting
