+++ src/plugins/nonebot_plugin_larkcave/commands/__init__.py
@@ -15,4 +15,4 @@
 #  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 # ##############################################################################
 
-from . import cave, get, post, remove, restore, statisics
+from . import cave, get, post, prompt, remove, restore, statisics
